issue533: added more dataupdateperiodenum value and separated function for better readability
